requirements.txt
dagster
dagster_duckdb_pandas
dagster_duckdb
matplotlib
pandas
requests
dagster-webserver
streamlit
supervisor